+++ example/android/app/src/main/kotlin/com/example/boolean_validation_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.boolean_validation_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()