+++ lib/boolean_validation.dart
@@ -1,0 +1,3 @@
+library txt_form_validation;
+export 'src/validators.dart';
+